@@ -15,10 +15,11 @@
         "sass": "^1.32.13",
         "sass-loader": "^7.3.1",
         "vite": "^3",
-        "vue": "^2.7.13",
+        "vue": "^3",
         "vue-template-compiler": "^2.6.14"
     },
     "dependencies": {
+        "@vitejs/plugin-vue": "^3.2.0",
         "acorn-dynamic-import": "^4.0.0",
         "axios": "^0.19.2",
         "css-unit-converter": "^1.1.2",
@@ -28,8 +29,6 @@
         "mixin-object": "^3.0.0",
         "select2": "^4.0.13",
         "v-tooltip": "^2.1.3",
-        "vite-plugin-require": "^1.0.7",
-        "vite-plugin-vue2": "^2.0.2",
-        "vue-d3-charts": "^0.2.8"
+        "vite-plugin-require": "^1.0.7"
     }
 }
