@@ -9,14 +9,14 @@
         >
             <strong>{{ name }}:</strong> {{ group.name }}
         </li>
-        <div 
+        <div></div>
     </ul>
     <div v-else-if="groupsLoaded && !Object.keys(groups).length"
         class="alert alert-warning">Du er ikke med i noen grupper. <p>For å være med i gruppediskusjoner må du velge din tilhørighet lenger ned på denne siden.</p>
     </div>
     <p v-else>
       Laster inn dine grupper...
-    </p>  
+    </p>
 
 </template>
 
