@@ -7,7 +7,7 @@
     <!-- Fonts -->
     <link rel="dns-prefetch" href="//fonts.gstatic.com">
     <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet" type="text/css">
-    @vite(['resources/sass/app.scss', 'resources/js/app.js'])
+    @vite('resources/sass/app.scss')
 
 </head>
 <body>
@@ -15,5 +15,6 @@
         @yield('content')
     </div>
     @yield('scripts')
+    @vite('resources/js/app.js')
 </body>
 </html>
