@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
-import { createVuePlugin as vue } from "vite-plugin-vue2";
+import vue from '@vitejs/plugin-vue';
 
 export default defineConfig({
     css: {
@@ -11,7 +11,7 @@ export default defineConfig({
                `
               @import "./resourses/sass/_variables.scss";
               @import "./resourses/sass/app.scss";
-              @import "./resourses/sass/statistics.scss"; 
+              @import "./resourses/sass/statistics.scss";
             `
             }
           }
